api/requirements.txt
Django==4.2.21
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.1.0 